requirements.txt
web3>=6.0.0,<7.0.0
python-dotenv>=1.0.0
rich>=13.0.0
